ai-service-python/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
sentence-transformers
faiss-cpu
numpy
scikit-learn
matplotlib